+++ src/widget/radio_button_widget.rs
@@ -39,6 +39,7 @@ pub struct RadioButtonWidget {
     unselected_widget: ImageWidget,
     inject_event: bool,
     widget_id: i32,
+    on_click: Option<Box<dyn FnMut (&mut RadioButtonWidget)>>,
 }
 
 impl RadioButtonWidget {
@@ -73,6 +74,18 @@ impl RadioButtonWidget {
             unselected_widget,
             inject_event: false,
             widget_id: 0,
+            on_click: None,
+        }
+    }
+
+    pub fn on_click<F>(&mut self, callback: F) where F: FnMut (&mut RadioButtonWidget) + 'static {
+        self.on_click = Some(Box::new(callback));
+    }
+
+    pub fn click(&mut self) {
+        if let Some(mut cb) = self.on_click.take() {
+            cb(self);
+            self.on_click = Some(cb);
         }
     }
 }
@@ -163,6 +176,7 @@ impl Widget for RadioButtonWidget {
                         if mouse_button == MouseButton::Left {
                             self.selected = true;
                             self.inject_event = true;
+                            self.click();
 
                             self.invalidate();